+++ packages/playground/vue/Main.vue
@@ -11,6 +11,10 @@
   <Assets />
   <CustomBlock />
   <SrcImport />
+  <Slotted>
+    <div class="slotted">this should be red</div>
+  </Slotted>
+  <ScanDep/>
 </template>
 
 <script setup lang="ts">
@@ -21,6 +25,8 @@ import CssModules from './CssModules.vue'
 import Assets from './Assets.vue'
 import CustomBlock from './CustomBlock.vue'
 import SrcImport from './src-import/SrcImport.vue'
+import Slotted from './Slotted.vue'
+import ScanDep from './ScanDep.vue'
 
 import { ref } from 'vue'