+++ packages/vite/src/node/plugins/css.ts
@@ -34,8 +34,10 @@ import {
 } from './asset'
 import MagicString from 'magic-string'
 import * as Postcss from 'postcss'
-import * as Sass from 'sass'
+import type Sass from 'sass'
+import type Stylus from 'stylus'
 import type Less from 'less'
+import { Alias } from 'types/alias'
 
 // const debug = createDebugger('vite:css')
 
@@ -80,6 +82,18 @@ const cssLangRE = new RegExp(cssLangs)
 const cssModuleRE = new RegExp(`\\.module${cssLangs}`)
 const directRequestRE = /(\?|&)direct\b/
 const commonjsProxyRE = /\?commonjs-proxy/
+
+const enum PreprocessLang {
+  less = 'less',
+  sass = 'sass',
+  scss = 'scss',
+  styl = 'styl',
+  stylus = 'stylus'
+}
+const enum PureCssLang {
+  css = 'css'
+}
+type CssLang = keyof typeof PureCssLang | keyof typeof PreprocessLang
 
 export const isCSSRequest = (request: string) =>
   cssLangRE.test(request) && !directRequestRE.test(request)
@@ -300,7 +314,7 @@ export function cssPostPlugin(config: ResolvedConfig): Plugin {
           const filename = getAssetFilename(fileHash, config) + postfix
           registerAssetToChunk(chunk, filename)
           if (!isRelativeBase || inlined) {
-            // absoulte base or relative base but inlined (injected as style tag into
+            // absolute base or relative base but inlined (injected as style tag into
             // index.html) use the base as-is
             return config.base + filename
           } else {
@@ -362,7 +376,7 @@ export function cssPostPlugin(config: ResolvedConfig): Plugin {
           }
         }
       } else {
-        // non-split extracted CSS will be minified togethter
+        // non-split extracted CSS will be minified together
         chunkCSS = await processChunkCSS(chunkCSS, {
           inlined: false,
           minify: false
@@ -508,7 +522,7 @@ async function compileCSS(
   const needInlineImport = code.includes('@import')
   const hasUrl = cssUrlRE.test(code) || cssImageSetRE.test(code)
   const postcssConfig = await resolvePostcssConfig(config)
-  const lang = id.match(cssLangRE)?.[1]
+  const lang = id.match(cssLangRE)?.[1] as CssLang | undefined
 
   // 1. plain css that needs no processing
   if (
@@ -526,23 +540,25 @@ async function compileCSS(
   const deps = new Set<string>()
 
   // 2. pre-processors: sass etc.
-  if (lang && lang in preProcessors) {
-    const preProcessor = preProcessors[lang as PreprocessLang]
+  if (isPreProcessor(lang)) {
+    const preProcessor = preProcessors[lang]
     let opts = (preprocessorOptions && preprocessorOptions[lang]) || {}
     // support @import from node dependencies by default
     switch (lang) {
-      case 'scss':
-      case 'sass':
+      case PreprocessLang.scss:
+      case PreprocessLang.sass:
         opts = {
           includePaths: ['node_modules'],
+          alias: config.resolve.alias,
           ...opts
         }
         break
-      case 'less':
-      case 'styl':
-      case 'stylus':
+      case PreprocessLang.less:
+      case PreprocessLang.styl:
+      case PreprocessLang.stylus:
         opts = {
           paths: ['node_modules'],
+          alias: config.resolve.alias,
           ...opts
         }
     }
@@ -717,19 +733,21 @@ const UrlRewritePostcssPlugin: Postcss.PluginCreator<{
     postcssPlugin: 'vite-url-rewrite',
     Once(root) {
       const promises: Promise<void>[] = []
-      root.walkDecls((decl) => {
-        const isCssUrl = cssUrlRE.test(decl.value)
-        const isCssImageSet = cssImageSetRE.test(decl.value)
+      root.walkDecls((declaration) => {
+        const isCssUrl = cssUrlRE.test(declaration.value)
+        const isCssImageSet = cssImageSetRE.test(declaration.value)
         if (isCssUrl || isCssImageSet) {
-          const replacerForDecl = (rawUrl: string) => {
-            const importer = decl.source?.input.file
+          const replacerForDeclaration = (rawUrl: string) => {
+            const importer = declaration.source?.input.file
             return opts.replacer(rawUrl, importer)
           }
           const rewriterToUse = isCssUrl ? rewriteCssUrls : rewriteCssImageSet
           promises.push(
-            rewriterToUse(decl.value, replacerForDecl).then((url) => {
-              decl.value = url
-            })
+            rewriterToUse(declaration.value, replacerForDeclaration).then(
+              (url) => {
+                declaration.value = url
+              }
+            )
           )
         }
       })
@@ -835,8 +853,6 @@ AtImportHoistPlugin.postcss = true
 
 // Preprocessor support. This logic is largely replicated from @vue/compiler-sfc
 
-type PreprocessLang = 'less' | 'sass' | 'scss' | 'styl' | 'stylus'
-
 type PreprocessorAdditionalData =
   | string
   | ((source: string, filename: string) => string | Promise<string>)
@@ -848,6 +864,7 @@ type StylePreprocessor = (
     [key: string]: any
     additionalData?: PreprocessorAdditionalData
     filename: string
+    alias: Alias[]
   },
   resolvers: CSSAtImportResolvers
 ) => StylePreprocessorResults | Promise<StylePreprocessorResults>
@@ -861,7 +878,14 @@ export interface StylePreprocessorResults {
 
 const loadedPreprocessors: Partial<Record<PreprocessLang, any>> = {}
 
-function loadPreprocessor(lang: PreprocessLang, root: string) {
+function loadPreprocessor(lang: PreprocessLang.scss, root: string): typeof Sass
+function loadPreprocessor(lang: PreprocessLang.sass, root: string): typeof Sass
+function loadPreprocessor(lang: PreprocessLang.less, root: string): typeof Less
+function loadPreprocessor(
+  lang: PreprocessLang.stylus,
+  root: string
+): typeof Stylus
+function loadPreprocessor(lang: PreprocessLang, root: string): any {
   if (lang in loadedPreprocessors) {
     return loadedPreprocessors[lang]
   }
@@ -877,7 +901,7 @@ function loadPreprocessor(lang: PreprocessLang, root: string) {
 
 // .scss/.sass processor
 const scss: StylePreprocessor = async (source, root, options, resolvers) => {
-  const render = loadPreprocessor('sass', root).render as typeof Sass.render
+  const render = loadPreprocessor(PreprocessLang.sass, root).render
   const finalOptions: Sass.Options = {
     ...options,
     data: await getSource(source, options.filename, options.additionalData),
@@ -886,7 +910,7 @@ const scss: StylePreprocessor = async (source, root, options, resolvers) => {
     importer(url, importer, done) {
       resolvers.sass(url, importer).then((resolved) => {
         if (resolved) {
-          rebaseUrls(resolved, options.filename).then(done)
+          rebaseUrls(resolved, options.filename, options.alias).then(done)
         } else {
           done(null)
         }
@@ -936,7 +960,8 @@ const sass: StylePreprocessor = (source, root, options, aliasResolver) =>
  */
 async function rebaseUrls(
   file: string,
-  rootFile: string
+  rootFile: string,
+  alias: Alias[]
 ): Promise<Sass.ImporterReturnType> {
   file = path.resolve(file) // ensure os-specific flashes
   // in the same dir, no need to rebase
@@ -952,6 +977,14 @@ async function rebaseUrls(
   }
   const rebased = await rewriteCssUrls(content, (url) => {
     if (url.startsWith('/')) return url
+    // match alias, no need to rewrite
+    for (const { find } of alias) {
+      const matches =
+        typeof find === 'string' ? url.startsWith(find) : find.test(url)
+      if (matches) {
+        return url
+      }
+    }
     const absolute = path.resolve(fileDir, url)
     const relative = path.relative(rootDir, absolute)
     return normalizePath(relative)
@@ -964,10 +997,11 @@ async function rebaseUrls(
 
 // .less
 const less: StylePreprocessor = async (source, root, options, resolvers) => {
-  const nodeLess = loadPreprocessor('less', root) as typeof Less
+  const nodeLess = loadPreprocessor(PreprocessLang.less, root)
   const viteResolverPlugin = createViteLessPlugin(
     nodeLess,
     options.filename,
+    options.alias,
     resolvers
   )
   source = await getSource(source, options.filename, options.additionalData)
@@ -1004,16 +1038,23 @@ let ViteLessManager: any
 function createViteLessPlugin(
   less: typeof Less,
   rootFile: string,
+  alias: Alias[],
   resolvers: CSSAtImportResolvers
 ): Less.Plugin {
   if (!ViteLessManager) {
     ViteLessManager = class ViteManager extends less.FileManager {
       resolvers
       rootFile
-      constructor(rootFile: string, resolvers: CSSAtImportResolvers) {
+      alias
+      constructor(
+        rootFile: string,
+        resolvers: CSSAtImportResolvers,
+        alias: Alias[]
+      ) {
         super()
         this.rootFile = rootFile
         this.resolvers = resolvers
+        this.alias = alias
       }
       supports() {
         return true
@@ -1032,7 +1073,7 @@ function createViteLessPlugin(
           path.join(dir, '*')
         )
         if (resolved) {
-          const result = await rebaseUrls(resolved, this.rootFile)
+          const result = await rebaseUrls(resolved, this.rootFile, this.alias)
           let contents
           if (result && 'contents' in result) {
             contents = result.contents
@@ -1052,7 +1093,9 @@ function createViteLessPlugin(
 
   return {
     install(_, pluginManager) {
-      pluginManager.addFileManager(new ViteLessManager(rootFile, resolvers))
+      pluginManager.addFileManager(
+        new ViteLessManager(rootFile, resolvers, alias)
+      )
     },
     minVersion: [3, 0, 0]
   }
@@ -1060,13 +1103,16 @@ function createViteLessPlugin(
 
 // .styl
 const styl: StylePreprocessor = (source, root, options) => {
-  const nodeStylus = loadPreprocessor('stylus', root)
+  const nodeStylus = loadPreprocessor(PreprocessLang.stylus, root)
   try {
     const ref = nodeStylus(source)
+
     Object.keys(options).forEach((key) => ref.set(key, options[key]))
     // if (map) ref.set('sourcemap', { inline: false, comment: false })
 
     const result = ref.render()
+
+    // @ts-expect-error: https://github.com/DefinitelyTyped/DefinitelyTyped/pull/51919
     const deps = ref.deps()
 
     return { code: result, errors: [], deps }
@@ -1087,10 +1133,14 @@ function getSource(
   return additionalData + source
 }
 
-const preProcessors = {
-  less,
-  sass,
-  scss,
-  styl,
-  stylus: styl
+const preProcessors = Object.freeze({
+  [PreprocessLang.less]: less,
+  [PreprocessLang.sass]: sass,
+  [PreprocessLang.scss]: scss,
+  [PreprocessLang.styl]: styl,
+  [PreprocessLang.stylus]: styl
+})
+
+function isPreProcessor(lang: any): lang is PreprocessLang {
+  return lang && lang in preProcessors
 }